Initial commit
Make a bower component for MIDImoji. Some of the `bower.json` is kind of
a lie at the moment, because midimoji.js doesn't exist. I'll fix that in
a sec.

I also got a bit carried away with figuring out how to identify a major
triad, rather than pursuing the stated aim of putting an emoji on the
page when a MIDI signal is received.

Oh well.